backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
djangorestframework-simplejwt==5.3.1
django-filter==23.5
python-dateutil==2.8.2
pytz==2024.1
# Jira Integration
jira==3.8.0
